# Rect `collidelist()`/`collidelistall()` speedup