--- a/src_c/rect.c
+++ b/src_c/rect.c
@@ -718,97 +718,128 @@ pg_rect_colliderect(pgRectObject *self, PyObject *const *args,
 PG_WRAP_FASTCALL_FUNC(pg_rect_colliderect, pgRectObject)
 
 static PyObject *
-pg_rect_collidelist(pgRectObject *self, PyObject *args)
+pg_rect_collidelist(pgRectObject *self, PyObject *arg)
 {
     SDL_Rect *argrect, temp;
-    Py_ssize_t size;
-    int loop;
-    PyObject *list, *obj;
-    PyObject *ret = NULL;
+    const SDL_Rect *const srect = &self->r;
+    Py_ssize_t loop;
 
-    if (!PyArg_ParseTuple(args, "O", &list)) {
-        return NULL;
-    }
-
-    if (!PySequence_Check(list)) {
+    if (!PySequence_Check(arg)) {
         return RAISE(PyExc_TypeError,
                      "Argument must be a sequence of rectstyle objects.");
     }
 
-    size = PySequence_Length(list); /*warning, size could be -1 on error?*/
-    for (loop = 0; loop < size; ++loop) {
-        obj = PySequence_GetItem(list, loop);
-        if (!obj || !(argrect = pgRect_FromObject(obj, &temp))) {
-            PyErr_SetString(
-                PyExc_TypeError,
-                "Argument must be a sequence of rectstyle objects.");
-            Py_XDECREF(obj);
-            break;
+    /* If the sequence is a fast sequence, we can use the faster
+     * PySequence_Fast_ITEMS() function to get the items. */
+    if (pgSequenceFast_Check(arg)) {
+        PyObject **items = PySequence_Fast_ITEMS(arg);
+        for (loop = 0; loop < PySequence_Fast_GET_SIZE(arg); loop++) {
+            if (!(argrect = pgRect_FromObject(items[loop], &temp))) {
+                return RAISE(
+                    PyExc_TypeError,
+                    "Argument must be a sequence of rectstyle objects.");
+            }
+            if (_pg_do_rects_intersect(srect, argrect)) {
+                return PyLong_FromLong(loop);
+            }
         }
-        if (_pg_do_rects_intersect(&self->r, argrect)) {
-            ret = PyLong_FromLong(loop);
-            Py_DECREF(obj);
-            break;
-        }
-        Py_DECREF(obj);
     }
-    if (loop == size) {
-        ret = PyLong_FromLong(-1);
+    /* If the sequence is not a fast sequence, we have to use the slower
+     * PySequence_GetItem() function to get the items. */
+    else {
+        for (loop = 0; loop < PySequence_Length(arg); loop++) {
+            PyObject *obj = PySequence_GetItem(arg, loop);
+            if (!obj || !(argrect = pgRect_FromObject(obj, &temp))) {
+                Py_XDECREF(obj);
+                return RAISE(
+                    PyExc_TypeError,
+                    "Argument must be a sequence of rectstyle objects.");
+            }
+            if (_pg_do_rects_intersect(srect, argrect)) {
+                Py_DECREF(obj);
+                return PyLong_FromLong(loop);
+            }
+            Py_DECREF(obj);
+        }
     }
 
-    return ret;
+    return PyLong_FromLong(-1);
 }
 
 static PyObject *
-pg_rect_collidelistall(pgRectObject *self, PyObject *args)
+pg_rect_collidelistall(pgRectObject *self, PyObject *arg)
 {
     SDL_Rect *argrect, temp;
-    Py_ssize_t size;
-    int loop;
-    PyObject *list, *obj;
+    const SDL_Rect *const srect = &self->r;
+    Py_ssize_t loop;
     PyObject *ret = NULL;
 
-    if (!PyArg_ParseTuple(args, "O", &list)) {
-        return NULL;
-    }
-
-    if (!PySequence_Check(list)) {
+    if (!PySequence_Check(arg)) {
         return RAISE(PyExc_TypeError,
                      "Argument must be a sequence of rectstyle objects.");
     }
 
-    ret = PyList_New(0);
-    if (!ret) {
+    if (!(ret = PyList_New(0))) {
         return NULL;
     }
 
-    size = PySequence_Length(list); /*warning, size could be -1?*/
-    for (loop = 0; loop < size; ++loop) {
-        obj = PySequence_GetItem(list, loop);
-
-        if (!obj || !(argrect = pgRect_FromObject(obj, &temp))) {
-            Py_XDECREF(obj);
-            Py_DECREF(ret);
-            return RAISE(PyExc_TypeError,
-                         "Argument must be a sequence of rectstyle objects.");
-        }
-
-        if (_pg_do_rects_intersect(&self->r, argrect)) {
-            PyObject *num = PyLong_FromLong(loop);
-            if (!num) {
+    /* If the sequence is a fast sequence, we can use the faster
+     * PySequence_Fast_ITEMS() function to get the items. */
+    if (pgSequenceFast_Check(arg)) {
+        PyObject **items = PySequence_Fast_ITEMS(arg);
+        for (loop = 0; loop < PySequence_Fast_GET_SIZE(arg); loop++) {
+            if (!(argrect = pgRect_FromObject(items[loop], &temp))) {
                 Py_DECREF(ret);
-                Py_DECREF(obj);
-                return NULL;
+                return RAISE(
+                    PyExc_TypeError,
+                    "Argument must be a sequence of rectstyle objects.");
             }
-            if (0 != PyList_Append(ret, num)) {
-                Py_DECREF(ret);
+
+            if (_pg_do_rects_intersect(srect, argrect)) {
+                PyObject *num = PyLong_FromLong(loop);
+                if (!num) {
+                    Py_DECREF(ret);
+                    return NULL;
+                }
+                if (PyList_Append(ret, num)) {
+                    Py_DECREF(ret);
+                    Py_DECREF(num);
+                    return NULL; /* Exception already set. */
+                }
                 Py_DECREF(num);
-                Py_DECREF(obj);
-                return NULL; /* Exception already set. */
             }
-            Py_DECREF(num);
         }
-        Py_DECREF(obj);
+    }
+    /* If the sequence is not a fast sequence, we have to use the slower
+     * PySequence_GetItem() function to get the items. */
+    else {
+        for (loop = 0; loop < PySequence_Length(arg); loop++) {
+            PyObject *obj = PySequence_GetItem(arg, loop);
+            if (!obj || !(argrect = pgRect_FromObject(obj, &temp))) {
+                Py_XDECREF(obj);
+                Py_DECREF(ret);
+                return RAISE(
+                    PyExc_TypeError,
+                    "Argument must be a sequence of rectstyle objects.");
+            }
+
+            if (_pg_do_rects_intersect(srect, argrect)) {
+                PyObject *num = PyLong_FromLong(loop);
+                if (!num) {
+                    Py_DECREF(ret);
+                    Py_DECREF(obj);
+                    return NULL;
+                }
+                if (PyList_Append(ret, num)) {
+                    Py_DECREF(ret);
+                    Py_DECREF(num);
+                    Py_DECREF(obj);
+                    return NULL; /* Exception already set. */
+                }
+                Py_DECREF(num);
+            }
+            Py_DECREF(obj);
+        }
     }
 
     return ret;
@@ -1401,9 +1432,9 @@ static struct PyMethodDef pg_rect_methods[] = {
      PG_FASTCALL, DOC_RECTCOLLIDEPOINT},
     {"colliderect", (PyCFunction)PG_FASTCALL_NAME(pg_rect_colliderect),
      PG_FASTCALL, DOC_RECTCOLLIDERECT},
-    {"collidelist", (PyCFunction)pg_rect_collidelist, METH_VARARGS,
+    {"collidelist", (PyCFunction)pg_rect_collidelist, METH_O,
      DOC_RECTCOLLIDELIST},
-    {"collidelistall", (PyCFunction)pg_rect_collidelistall, METH_VARARGS,
+    {"collidelistall", (PyCFunction)pg_rect_collidelistall, METH_O,
      DOC_RECTCOLLIDELISTALL},
     {"collideobjectsall", (PyCFunction)pg_rect_collideobjectsall,
      METH_VARARGS | METH_KEYWORDS, DOC_RECTCOLLIDEOBJECTSALL},